requirements.txt
aiodns
python-whois
httpx
beautifulsoup4
dnspython
requests
ipaddress
xmltodict
